🌙 Add better dark mode
-Refactor AuthLayout.tsx
