+++ resources/js/src/Utils/Utils.ts
@@ -1,5 +1,5 @@
 import {UseFormReturn} from "react-hook-form";
-import {usePage} from "@inertiajs/react";
+import {router, usePage} from "@inertiajs/react";
 import React from "react";
 
 const mapServerErrorsToFields = function (form: UseFormReturn, errors: Record<string, string>) {
@@ -68,4 +68,27 @@ const useOnlineStatus = () => {
     return isOnline;
 }
 
-export {usePermission, useOnlineStatus, mergeDeep, isObject, mapServerErrorsToFields}
+const useLoading = () => {
+
+    const [loading, setLoading] = React.useState(false)
+
+    React.useEffect(() => {
+        let timeout = null
+        const onStartListener = router.on('start', () => {
+            timeout = setTimeout(() => setLoading(true), 250)
+        })
+        const onFinishListener = router.on('finish', () => {
+            clearTimeout(timeout)
+            setLoading(false)
+        })
+
+        return () => {
+            onStartListener()
+            onFinishListener()
+        }
+    }, [])
+
+    return loading
+}
+
+export {usePermission, useOnlineStatus, useLoading, mergeDeep, isObject, mapServerErrorsToFields}
